requirement.txt
streamlit
langchain
langchain-community
openai
python-dotenv
pypdf
pyPDF2
faiss-cpu
streamlit-extras